refactor: move all files related to UI(activities, etc.) into package 'ui'
