+++ app/src/main/AndroidManifest.xml
@@ -9,13 +9,13 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".preferences.PreferencesActivity" />
-        <activity android:name=".mypage.EditProfileActivity" />
-        <activity android:name=".mypage.MyPageActivity" />
-        <activity android:name=".mypage.CreateProfileActivity" />
-        <activity android:name=".NotifyAgreementActivity" />
-        <activity android:name=".signup.SignUpActivity" />
-        <activity android:name=".MainActivity">
+        <activity android:name=".ui.preferences.PreferencesActivity" />
+        <activity android:name=".ui.mypage.EditProfileActivity" />
+        <activity android:name=".ui.mypage.MyPageActivity" />
+        <activity android:name=".ui.mypage.CreateProfileActivity" />
+        <activity android:name=".ui.NotifyAgreementActivity" />
+        <activity android:name=".ui.signup.SignUpActivity" />
+        <activity android:name=".ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
