+++ app/src/main/java/com/mashup/torchlight/ui/mypage/CreateProfileActivity.kt
@@ -1,13 +1,13 @@
-package com.mashup.torchlight.mypage
+package com.mashup.torchlight.ui.mypage
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.mashup.torchlight.R
 
-class MyPageActivity : AppCompatActivity() {
+class CreateProfileActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_my_page)
+        setContentView(R.layout.activity_create_profile)
     }
 }
