+++ app/src/main/java/com/mashup/torchlight/ui/mypage/CreateProfileSuccessFragment.kt
@@ -1,4 +1,4 @@
-package com.mashup.torchlight.mypage
+package com.mashup.torchlight.ui.mypage
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,7 +8,7 @@ import android.view.ViewGroup
 
 import com.mashup.torchlight.R
 
-class CreateProfileConfirmFragment : Fragment() {
+class CreateProfileSuccessFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -20,14 +20,15 @@ class CreateProfileConfirmFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_create_profile_confirm, container, false)
+        return inflater.inflate(R.layout.fragment_create_profile_success, container, false)
     }
 
 
     companion object {
+
         @JvmStatic
         fun newInstance() =
-            CreateProfileConfirmFragment().apply {
+            CreateProfileSuccessFragment().apply {
 
             }
     }
