+++ app/src/main/java/com/mashup/torchlight/ui/mypage/EditProfileOptionalInfoFragment.kt
@@ -1,4 +1,4 @@
-package com.mashup.torchlight.mypage
+package com.mashup.torchlight.ui.mypage
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,7 +8,7 @@ import android.view.ViewGroup
 
 import com.mashup.torchlight.R
 
-class CreateProfileSuccessFragment : Fragment() {
+class EditProfileOptionalInfoFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -20,15 +20,13 @@ class CreateProfileSuccessFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_create_profile_success, container, false)
+        return inflater.inflate(R.layout.fragment_create_profile_optional_info, container, false)
     }
 
-
     companion object {
-
         @JvmStatic
         fun newInstance() =
-            CreateProfileSuccessFragment().apply {
+            EditProfileOptionalInfoFragment().apply {
 
             }
     }
