+++ app/src/main/java/com/mashup/torchlight/ui/mypage/MyPageActivity.kt
@@ -1,13 +1,13 @@
-package com.mashup.torchlight.preferences
+package com.mashup.torchlight.ui.mypage
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.mashup.torchlight.R
 
-class PreferencesActivity : AppCompatActivity() {
+class MyPageActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_preferences)
+        setContentView(R.layout.activity_my_page)
     }
 }
