+++ app/src/main/java/com/mashup/torchlight/ui/preferences/PreferencesActivity.kt
@@ -1,13 +1,13 @@
-package com.mashup.torchlight.signup
+package com.mashup.torchlight.ui.preferences
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.mashup.torchlight.R
 
-class SignUpActivity : AppCompatActivity() {
+class PreferencesActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_sign_up)
+        setContentView(R.layout.activity_preferences)
     }
 }
