+++ app/src/main/java/com/mashup/torchlight/ui/signup/SignUpSuccessFragment.kt
@@ -1,4 +1,4 @@
-package com.mashup.torchlight.signup
+package com.mashup.torchlight.ui.signup
 
 
 import android.os.Bundle
