+++ app/src/main/java/com/mashup/torchlight/ui/signup/SignUpUserInfoFragment.kt
@@ -1,4 +1,4 @@
-package com.mashup.torchlight.signup
+package com.mashup.torchlight.ui.signup
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,7 +8,7 @@ import android.view.ViewGroup
 
 import com.mashup.torchlight.R
 
-class SignUpEmailAuthFragment : Fragment() {
+class SignUpUserInfoFragment : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -20,15 +20,14 @@ class SignUpEmailAuthFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_sign_up_email_auth_input, container, false)
+        return inflater.inflate(R.layout.fragment_sign_up_user_info, container, false)
     }
 
     companion object {
         @JvmStatic
         fun newInstance() =
-            SignUpEmailAuthFragment().apply {
+            SignUpUserInfoFragment().apply {
 
             }
     }
 }
-
