+++ app/src/main/java/com/mashup/torchlight/ui/signup/SignUpViewModel.kt
@@ -1,4 +1,4 @@
-package com.mashup.torchlight.signup
+package com.mashup.torchlight.ui.signup
 
 import androidx.lifecycle.ViewModel
