+++ app/src/main/res/layout/activity_create_profile.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".mypage.CreateProfileActivity">
+    tools:context=".ui.mypage.CreateProfileActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
