+++ app/src/main/res/layout/activity_my_page.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".mypage.MyPageActivity">
+    tools:context=".ui.mypage.MyPageActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
