+++ app/src/main/res/layout/activity_notify_agreement.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NotifyAgreementActivity">
+    tools:context=".ui.NotifyAgreementActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
