+++ app/src/main/res/layout/fragment_create_profile_confirm.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".mypage.CreateProfileConfirmFragment">
+    tools:context=".ui.mypage.CreateProfileConfirmFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
