+++ app/src/main/res/layout/fragment_create_profile_optional_info.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".mypage.EditProfileOptionalInfoFragment">
+    tools:context=".ui.mypage.EditProfileOptionalInfoFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
