+++ app/src/main/res/layout/fragment_create_profile_success.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".mypage.CreateProfileSuccessFragment">
+    tools:context=".ui.mypage.CreateProfileSuccessFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
