+++ app/src/main/res/layout/fragment_sign_up_email_auth_input.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".signup.SignUpEmailAuthFragment">
+    tools:context=".ui.signup.SignUpEmailAuthFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
