+++ app/src/main/res/layout/sign_up_email_auth_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".signup.SignUpEmailInputFragment">
+    tools:context=".ui.signup.SignUpEmailInputFragment">
 
     <TextView
         android:layout_width="match_parent"
